Actually Implementing Clipping.
Clipping earlier was not working. At all.

It works now.

There are a few bugs though.

All clipping is geometric clipping, as in, triangles are generated to fill in gaps that would be left from clipping triangles that lie out of the Homogeneous space and the NDC space.

Z clipping is done in world space. (Basically Frustum culling.)
X,Y culling is done in NDC space.