Testing UI.
Small tests to see what api changes need to be made to the ui script so far.

UI needs an initialization stage and several other things can be cleaned up.

Tests on moving UI still need to be done.

One possible bug, due to how UIstates are calculated is that UI Rects outside the bounds of their parent rects will not properly calculate hoverEnd events.

Is this a feature or a bug....probably a feature since currently my use case does not warrant taking into account such a condition.